getOutput: return the contents of a transform
The returnOutput() function is badly named, since it actually prints
the transform results to stdio rather than returning the results to
the caller. To preserve API compatibility a new function is introduced
which returns a list of dicts of entities for further processing.

This also facilitates the creation to scripts which can either be run
as command-line tools or as local transforms depending on switches or
environment.